misc: Move types from helper file to `/types` and create transformer.
