+++ types/discordeno.ts
@@ -1,4 +1,13 @@
-import { AllowedMentionsTypes, ButtonStyles, MessageComponentTypes, TextStyles } from "./shared.ts";
+import { ApplicationCommandOption } from "../transformers/applicationCommandOption.ts";
+import {
+  AllowedMentionsTypes,
+  ApplicationCommandTypes,
+  ButtonStyles,
+  Localization,
+  MessageComponentTypes,
+  PermissionStrings,
+  TextStyles,
+} from "./shared.ts";
 
 export type MessageComponents = ActionRow[];
 
@@ -126,4 +135,45 @@ export interface SearchMembers {
   query: string;
   /** Max number of members to return (1-1000). Default: 1 */
   limit?: number;
+}
+
+export type CreateApplicationCommand = CreateSlashApplicationCommand | CreateContextApplicationCommand;
+
+/** https://discord.com/developers/docs/interactions/application-commands#endpoints-json-params */
+export interface CreateSlashApplicationCommand {
+  /**
+   * Name of command, 1-32 characters.
+   * `ApplicationCommandTypes.ChatInput` command names must match the following regex `^[-_\p{L}\p{N}\p{sc=Deva}\p{sc=Thai}]{1,32}$` with the unicode flag set.
+   * If there is a lowercase variant of any letters used, you must use those.
+   * Characters with no lowercase variants and/or uncased letters are still allowed.
+   * ApplicationCommandTypes.User` and `ApplicationCommandTypes.Message` commands may be mixed case and can include spaces.
+   */
+  name: string;
+  /** Localization object for the `name` field. Values follow the same restrictions as `name` */
+  nameLocalizations?: Localization;
+  /** 1-100 character description */
+  description: string;
+  /** Localization object for the `description` field. Values follow the same restrictions as `description` */
+  descriptionLocalizations?: Localization;
+  /** Type of command, defaults `ApplicationCommandTypes.ChatInput` if not set  */
+  type?: ApplicationCommandTypes;
+  /** Parameters for the command */
+  options?: ApplicationCommandOption[];
+  /** Set of permissions represented as a bit set */
+  defaultMemberPermissions?: PermissionStrings[];
+  /** Indicates whether the command is available in DMs with the app, only for globally-scoped commands. By default, commands are visible. */
+  dmPermission?: boolean;
+}
+
+/** https://discord.com/developers/docs/interactions/application-commands#endpoints-json-params */
+export interface CreateContextApplicationCommand
+  extends Omit<CreateSlashApplicationCommand, "options" | "description" | "descriptionLocalizations"> {
+  /** The type of the command */
+  type: ApplicationCommandTypes.Message | ApplicationCommandTypes.User;
+}
+
+export function isContextApplicationCommand(
+  command: CreateApplicationCommand,
+): command is CreateContextApplicationCommand {
+  return command.type === ApplicationCommandTypes.Message || command.type === ApplicationCommandTypes.User;
 }
